chore: harden shared/ boundary — extract hashValues, drop apiKey, enforce model compat
